Fix word wrap issues caused by inline tags.
This fix introduces monitoring of the character count for the text line we're working with.  This is achieved by using a param on the options object that gets passed through the walker.  We update the line width value before processing the next element in the current dom.  The word-wrap function can then use this to work out where to break the text into lines.  Custom handling was needed for anchor tags as it was resetting the character count.

This also includes some new tests and a change to the expected output text as it contained lines that overran the 80 character limit.